#3 refactor: component 패키지 경로 수정
